Setup package format, add dev script
