+++ scripts/dev.bat
@@ -1,0 +1,4 @@
+@REM Sets up the repgen5 environment and loads the package in development mode
+
+python -m pip install -r requirements.txt
+python -m pip install -e .
